[chore] materialized_table: refactor test helpers for reuse
